+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 -   Open Jira issue image attachments within VS Code
 -   Support commit-level diffs for pull requests
 -   Add missing clone config for steps in pipelines yaml validator
+-   Atlassian Telemetry now respects telemetry.enableTelemetry flag in global settings
 
 ### Bugs fixed
 
 -   Fixed summary editor size on the create pull request screen
+-   Fixed styling for expander headers
 
 ## What's New in 2.8.3